+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # What's New with Enterprise-NG
 
+## 18.4.0
+
+## 18.4.0 Fixes
+
+- `[General]` All libraries (jQuery, Enterprise, Angular, D3) are now set to peer dependencies. We now recommend this. ([#4917](https://inforwiki.atlassian.net/browse/IDS-4917))
+
 ## 18.3.0
 
 ## 18.3.0 Features